refactor: Moved utility scripts to their own module
